+++ bld/mad/ppc/c/ppcdisas.c
@@ -70,7 +70,7 @@ dis_return DisCliGetData( void *d, unsigned off, unsigned size, void *data )
     return( DR_OK );
 }
 
-unsigned DisCliValueString( void *d, dis_dec_ins *ins, unsigned op, char *buff )
+size_t DisCliValueString( void *d, dis_dec_ins *ins, unsigned op, char *buff )
 {
     mad_disasm_data     *dd = d;
     char                *p;
